requirements.txt
streamlit>=1.30,<2.0
streamlit-folium>=0.10,<1.0
pydeck>=0.8,<1.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
pyarrow>=17.0,<19.0
requests>=2.31,<3.0
kafka-python>=2.0,<3.0
# confluent-kafka>=2.3,<3.0

# Phase 1: Spatial Intelligence
pyspark>=3.4.1,<3.5.0
apache-sedona==1.7.0
geopandas>=0.14.0
shapely>=2.0.0

# Sprint 6: Alert Sink Layer
duckdb>=1.0,<2.0
